Updated sockets.py default backlog
In CPython 3.5, the "backlog" parameter was made optional:
https://docs.python.org/3.5/library/socket.html#socket.socket.listen

After some diffing in the CPython source, I discovered that CPython uses 128 default backlog connections:
https://github.com/python/cpython/blob/2d264235f6e066611b412f7c2e1603866e0f7f1b/Modules/socketmodule.c#L2818

For users that are utilizing circuits for the first time, emulating this behaviour is helpful in ensuring a consistent developer experience. 
Why? Because 5k backlogged connections is considered obscenely large and will consume system resources very rapidly.